requirements/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-silk==5.3.2
django-simple-captcha==0.6.1
django-simple-history==3.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gprof2dot==2025.4.14
gunicorn==23.0.0
idna==3.10
inflect==7.5.0
more-itertools==10.6.0
# mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pgeocode==0.5.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
PyJWT==2.9.0
# PyMySQL==1.1.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
xlrd==2.0.1
